snapm: introduce snapshot set level locking in Manager
Add locking to Manager operations that modify external resources:
before operating on a named snapshot set, acquire a BSD-style flock()
lock on /run/lock/snapm/<name>.lock, and release the lock at the end of
the operation.

Initially this is a PoC/trial to evaluate the changes necessary and to
test them in CI. Depending on the results of the trial this may or may
not be included in a future release.

Resolves: #170